Fixed vector out of range issue
When re-selecting a fighter, s was assigned the same ID as the previous ID. If the user selected a fighter before the currently selected fighter in the vector, a vector out of range error occurred.  Changing the value of s to 0 every time the user enters pick a fighter fixes this.